Improved efficiency

+++ main.py
@@ -71,10 +71,16 @@ dp.process_handling = {
 
 # Sliders for filter parameters
 st.sidebar.header("Filter Settings (`-1` = disable)")
+# chi_min = st.sidebar.slider("Reduced chi^2 min", 0, 100, 0)
+# chi_max = st.sidebar.slider("Reduced chi^2 max", -1, 1000, 50)
+# km_min = st.sidebar.slider("Altitude (km) min", 0, 100, 0)
+# km_max = st.sidebar.slider("Altitude (km) max", -1, 100, 20)
+# mask_count_min = st.sidebar.slider("Mask minimum occurances", 1, 10, 2)
+
 chi_min = st.sidebar.slider("Reduced chi^2 min", 0, 100, 0)
 chi_max = st.sidebar.slider("Reduced chi^2 max", -1, 1000, 50)
-km_min = st.sidebar.slider("Altitude (km) min", 0, 100, 20)
-km_max = st.sidebar.slider("Altitude (km) max", -1, 1000, -1)
+km_min = st.sidebar.slider("Altitude (km) min", 0, 100, 0)
+km_max = st.sidebar.slider("Altitude (km) max", -1, 200, 200)
 mask_count_min = st.sidebar.slider("Mask minimum occurances", 1, 10, 2)
 
 do_topography_mapping = st.sidebar.checkbox(label="Enable Topography", value=False)
@@ -84,14 +90,11 @@ dp.downsampling_factor = st.sidebar.slider("Topography Downsampling (Size Reduct
 # Update the dp.filters and dp.count_required dynamically
 dp.filters = [
     ["reduced chi^2", lambda chi: chi >= chi_min],
-    ["alt(m)", lambda alt: alt >= km_min * 1000],
+    ["alt(m)", lambda alt: km_max * 1000 >= alt >= km_min * 1000],
 ]
 
 if chi_max >= 0:
     dp.filters.append(["reduced chi^2", lambda chi: chi <= chi_max])
-
-if km_max > 0:
-    dp.filters.append(["alt(m)", lambda alt: alt <= km_max * 1000])
 
 dp.count_required = [
     ["mask", lambda count: count >= mask_count_min]
@@ -130,54 +133,37 @@ def main():
         # Parse the selected file
         data_result: pd.DataFrame = dp.get_dataframe(lightning_data_folder, selected_file)
 
-        lowest_lon = 360
-        largest_lon = -360
-        for longitude in data_result['lon']:
-            if longitude > largest_lon:
-                largest_lon = longitude
-            if longitude < lowest_lon:
-                lowest_lon = longitude
-        
-        dp.params["west"] = lowest_lon - 1
-        dp.params["east"] = largest_lon + 1
+        if len(data_result['mask']) == 0:
+            st.write("No data matching the parameters")
 
-        lowest_lat = 360
-        highest_lat = -360
-        for latitude in data_result['lat']:
-            if latitude > highest_lat:
-                highest_lat = latitude
-            if latitude < lowest_lat:
-                lowest_lat = latitude
+        else:
 
-        dp.params["south"] = lowest_lat - 1
-        dp.params["north"] = highest_lat + 1
+            with st.spinner('Indexing Topography Data...'):
+                # Plot 3D scatter
+                fig = dp.plot_interactive_3d(data_result, 'mask', do_topography=do_topography_mapping)
 
-        with st.spinner('Indexing Topography Data...'):
-            # Plot 3D scatter
-            fig = dp.plot_interactive_3d(data_result, 'mask', do_topography=do_topography_mapping)
+            # Display the 3D plot in Streamlit
+            col1.plotly_chart(fig)
 
-        # Display the 3D plot in Streamlit
-        col1.plotly_chart(fig)
+            # Display the parsed DataFrame
+            col2.write("Parsed Data:")
+            col2.dataframe(dp.color_df(data_result, 'mask'))
+            
+            # Option to download the DataFrame as a CSV
+            csv_output = data_result.to_csv(index=False)
+            st.download_button(
+                label="Download CSV",
+                data=csv_output,
+                file_name=f"{os.path.splitext(selected_file)[0]}.csv",
+                mime="text/csv"
+            )
 
-        # Display the parsed DataFrame
-        col2.write("Parsed Data:")
-        col2.dataframe(dp.color_df(data_result, 'mask'))
-        
-        # Option to download the DataFrame as a CSV
-        csv_output = data_result.to_csv(index=False)
-        st.download_button(
-            label="Download CSV",
-            data=csv_output,
-            file_name=f"{os.path.splitext(selected_file)[0]}.csv",
-            mime="text/csv"
-        )
-
-        # Option to delete the selected file
-        if st.button(f"Delete {selected_file}"):
-            if delete_file(selected_file):
-                st.success(f"File '{selected_file}' deleted successfully!")
-            else:
-                st.error(f"Error: Could not delete file '{selected_file}'.")
+            # Option to delete the selected file
+            if st.button(f"Delete {selected_file}"):
+                if delete_file(selected_file):
+                    st.success(f"File '{selected_file}' deleted successfully!")
+                else:
+                    st.error(f"Error: Could not delete file '{selected_file}'.")
 
 
 ######################################################################################################
